requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.4
Django==4.1.3
django-google-maps==0.13.0
django-map-widgets==0.4.0
django-mapbox-location-field==2.0.0
docutils==0.19
pycparser==2.21
PyMySQL==1.0.2
sqlparse==0.4.3
tzdata==2022.6
